requirements.txt
art==4.2
awsebcli==3.15.2
blessed==1.15.0
boto3==1.9.183
botocore==1.12.183
cached-property==1.5.1
cachetools==3.1.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
coverage==4.5.4
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gspread==3.1.0
httplib2==0.13.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
oauth2client==4.1.3
oauthlib==3.0.2
pathspec==0.5.9
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
PyYAML==3.13
r7insight-python==1.0.0
requests==2.20.1
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==16.6.0
wcwidth==0.1.7
websocket-client==0.56.0
